libstagefright: Free buffers when observer dies
In case the observer dies from allocating the buffer that component remains
in a load state. It checks for idle state to free the buffer which obviously leads to a memory leak, that's why I explicitly call the free buffer to avoid that.

Perhaps acquiring mLock outside freeActivateBuffers() is a safe route.

Change-Id: I1b751d30fdb404029b96b605aa03c6b9ba8333ff